<refactoring> 바뀐 PostType 경로 지정
