+++ front/src/components/posts/card/LoadMorePostCards.tsx
@@ -78,23 +78,23 @@ const LoadMorePostCards: React.FC = () => {
             </Text>
             <PostSlider sliderName="User Quests" gap="20px" hight="440px">
                 {posts.map((post, index) => {
-                    if (index < 4)
-                        return (
-                            <MotionDiv
-                                variants={variants}
-                                initial="hidden"
-                                animate="visible"
-                                exit="hidden"
-                                transition={{
-                                    delay: index * 0.1,
-                                    ease: 'easeInOut',
-                                    duration: 0.5,
-                                }}
-                                viewport={{ amount: 0 }}
-                            >
-                                <PostCard post={post} key={post.id} index={index} type="quest" />
-                            </MotionDiv>
-                        );
+                    return (
+                        <MotionDiv
+                            key={post.id}
+                            variants={variants}
+                            initial="hidden"
+                            animate="visible"
+                            exit="hidden"
+                            transition={{
+                                delay: index * 0.1,
+                                ease: 'easeInOut',
+                                duration: 0.5,
+                            }}
+                            viewport={{ amount: 0 }}
+                        >
+                            <PostCard post={post} key={post.id} index={index} type="quest" />
+                        </MotionDiv>
+                    );
                 })}
             </PostSlider>
             <Box>
@@ -102,27 +102,6 @@ const LoadMorePostCards: React.FC = () => {
                     Submissions
                 </Text>
             </Box>
-            <PostSlider sliderName="Submission" gap="20px" hight="440px">
-                {posts.map((post, index) => {
-                    if (index < 4)
-                        return (
-                            <MotionDiv
-                                variants={variants}
-                                initial="hidden"
-                                animate="visible"
-                                exit="hidden"
-                                transition={{
-                                    delay: index * 0.1,
-                                    ease: 'easeInOut',
-                                    duration: 0.5,
-                                }}
-                                viewport={{ amount: 0 }}
-                            >
-                                <PostCard post={post} key={post.id} index={index} type="sb" />
-                            </MotionDiv>
-                        );
-                })}
-            </PostSlider>
             <section>
                 <span ref={ref}>
                     {isLoading && (
