+++ front/src/components/posts/card/MiniPostCard.tsx
@@ -3,8 +3,7 @@ import { Box, Flex, Text, Image, Spacer } from '@chakra-ui/react';
 import { customColors } from '@/utils/chakra/customColors';
 import React, { useCallback } from 'react';
 import { useRouter } from 'next/navigation';
-
-import { PostType } from '@/atoms/post';
+import { PostType } from '@/type/postType';
 
 type MiniPostCardProps = {
     post: PostType;
